changing int for size_t

+++ module04/ex01/Brain.cpp
@@ -6,7 +6,7 @@
 /*   By: ozamora- <ozamora-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/13 16:29:03 by ozamora-          #+#    #+#             */
-/*   Updated: 2025/09/15 15:43:20 by ozamora-         ###   ########.fr       */
+/*   Updated: 2025/09/15 16:42:39 by ozamora-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 Brain::Brain(void) {
 	std::cout << "Brain Default Constructor called\n";
-	for (int i = 0; i < 100 ; i++)
+	for (size_t i = 0; i < 100 ; i++)
 		ideas[i] = "";
 }
 
 Brain::Brain(const Brain& src) {
 	std::cout << "Brain Copy Constructor called\n";
-	for (int i = 0; i < 100; i++)
+	for (size_t i = 0; i < 100; i++)
 		ideas[i] = src.ideas[i];
 }
 
@@ -31,17 +31,17 @@ Brain::~Brain() {
 Brain& Brain::operator=(const Brain& src) {
 	std::cout << "Brain Copy Assignment Operator called\n";
 	if (this != &src) {
-		for (int i = 0; i < 100; i++)
+		for (size_t i = 0; i < 100; i++)
 			ideas[i] = src.ideas[i];
 	}
 	return *this;
 }
 
-std::string Brain::getIdea(int index) const {return ideas[index]; }
+std::string Brain::getIdea(size_t index) const {return ideas[index]; }
 
-void Brain::setIdea(int index, const std::string& idea) { ideas[index] = idea; }
+void Brain::setIdea(size_t index, const std::string& idea) { ideas[index] = idea; }
 
-void Brain::printIdea(int index) {
+void Brain::printIdea(size_t index) {
 	if (!ideas[index].empty())
 		std::cout << "Idea " << index << ": " << ideas[index] << std::endl;
 	else
@@ -49,7 +49,7 @@ void Brain::printIdea(int index) {
 }
 
 void Brain::printAllIdeas() {
-	for(int index = 0; index < 100; index++) {
+	for(size_t index = 0; index < 100; index++) {
 		if (!ideas[index].empty()) {
 			std::cout << "Idea " << index << ": " << ideas[index] << std::endl;
 		}
